Get enriched functional terms in gene families of a PLAZA instance

Get enriched functional terms in gene families of a PLAZA instance

get_family_functions(
  instance = c("Dicots", "Monocots"),
  type = c("homo", "ortho")
)

Arguments

instance

Character scalar with name of PLAZA instance for which gene family functions will be extracted. One of 'Dicots' or 'Monocots' (not available for 'Diatoms' and 'Pico').

type

Character indicating the type of gene family to use. One of 'homo' (homologous gene families) or 'ortho' (orthologous gene families).

Value

A data frame (or list of data frames, in case of multiple species) with the following variables:

family_id

Character, gene family ID.

term_id

Character, functional term ID.

term_description

Character, functional term description.

ontology

Character, ontology from where term comes.

score

Numeric, enrichment score.

Examples

fam_func <- get_family_functions("Dicots", type = "ortho")